Stworzyłem repozytoria, serisy oraz kontrollery, zmieniłem int id na long id 25.08 - 21:39

+++ Backend/projekt_inzynierski/src/main/java/models/Kraj_pochodzenia.java
@@ -21,7 +21,7 @@ public class Kraj_pochodzenia {
     @NotNull
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id_Kraj;
+    private long id_Kraj;
 
     @NotBlank
     @Size(max = 120)
